requirements.txt
fastapi==0.70.1
jinja2==3.0.3
python-dotenv==0.19.2
starlette==0.16.0
boto3==1.34.109
uvicorn==0.15.0
requests==2.32.3
httpx==0.27.0